0 remaining circular dependencies
