+++ package.json
@@ -33,8 +33,8 @@
     "test-e2e-relay": "cross-env MOCHA_SUITE_NAME=\"Mocha E2E Relay Tests\" c8 --report-dir='coverage/e2e-relay' mocha 'test/e2e/commands/relay.test.ts' --reporter-options configFile=mocha-multi-reporter.json,cmrOutput=mocha-junit-reporter+mochaFile+junit-e2e-relay.xml",
     "solo-test": "tsx --no-deprecation --no-warnings solo.ts",
     "solo": "node --no-deprecation --no-warnings dist/solo.js",
-    "check": "remark . --quiet --frail && eslint . && tsc && (madge --circular src/* || true) && cd docs; jsdoc -c jsdoc.conf.json",
-    "format": "remark . --quiet --frail --output && eslint --fix . && tsc && madge --circular src/* || true",
+    "check": "remark . --quiet --frail && eslint . && tsc && madge --circular src/* && cd docs; jsdoc -c jsdoc.conf.json",
+    "format": "remark . --quiet --frail --output && eslint --fix . && tsc && madge --circular src/*",
     "test-setup": "./test/e2e/setup-e2e.sh",
     "build": "rm -Rf dist && tsc && node resources/post-build-script.js"
   },
