+++ src/commands/base.ts
@@ -28,7 +28,7 @@ import {type ConfigManager} from '../core/config_manager.js';
 import {type DependencyManager} from '../core/dependency_managers/index.js';
 import {type Opts} from '../types/command_types.js';
 import {type CommandFlag} from '../types/flag_types.js';
-import {type Lease} from '../core/lease/types.js';
+import {type Lease} from '../core/lease/lease.js';
 import {Listr} from 'listr2';
 
 export interface CommandHandlers {
