+++ src/commands/node/tasks.ts
@@ -79,7 +79,7 @@ import type {
   NodeRefreshConfigClass,
   NodeUpdateConfigClass,
 } from './configs.js';
-import {type Lease} from '../../core/lease/types.js';
+import {type Lease} from '../../core/lease/lease.js';
 import {ListrLease} from '../../core/lease/listr_lease.js';
 import {Duration} from '../../core/time/duration.js';
 import {type BaseCommand} from '../base.js';
