+++ src/core/lease/listr_lease.ts
@@ -16,7 +16,7 @@
  */
 import {type ListrTaskWrapper} from 'listr2';
 import chalk from 'chalk';
-import {type Lease} from './types.js';
+import {type Lease} from './lease.js';
 import {LeaseAcquisitionError} from './lease_errors.js';
 
 /**
