+++ src/index.ts
@@ -43,8 +43,8 @@ import * as helpers from './core/helpers.js';
 import {K8} from './core/k8.js';
 import {CustomProcessOutput} from './core/process_output.js';
 import {type Opts} from './types/command_types.js';
-import {IntervalLeaseRenewalService} from './core/lease/lease_renewal.js';
-import {type LeaseRenewalService} from './core/lease/types.js';
+import {IntervalLeaseRenewalService} from './core/lease/interval_lease_renewal.js';
+import {type LeaseRenewalService} from './core/lease/lease.js';
 
 export function main(argv: any) {
   const logger = logging.NewLogger('debug');
